11101 Move Payments Menu rows to separate files
